Change jump function to accept non-existing matrices.
This change is necessary if we do not want to compute all matrices.
Now, if we cannot find a matrix to jump to some level l, we jump to
the smallest level l' with l' > l such that a matrix for l' exits and
then iterate until we reach the desired level.