insert parameter to control the number of matrices to compute.
This commit inserts a new parameter -j that accounts for the distance
between levels that are considered to be valid jump targets.

The main change is in the init_reach function that has to decide
which matrices to compute.

The jump function was already changed in a previous commit to deal
with "missing" matrices by doing several jumps if needed.