fix a crash in col_mul.
col mul might crash if the most significant bits in the column
vector are not set and therefore the underlying data_storage is
too short.

this commit also changes how the size of the matrix data structure
is stored. Instead of storing the padded_width, we store the number
of machine words.